==> part1_preprocessing.ipynb <==
# Part1 Pre-Processing

import pyspark
from pyspark.sql import SparkSession
from pyspark.sql.functions import *

spark = SparkSession.builder.appName("part1").getOrCreate()

# Bussiness

path = 'yelp_academic_dataset_business.json'
business = spark.read.json(path)
business.count()

business_col = ['business_id','name','city','state','stars','review_count','categories','latitude','longitude','is_open', 'postal_code']
business = business.select(business_col)
#business = business.withColumn('category', split(business['categories'],',')).withColumn('category',explode('category')).withColumn('category', trim('category'))


tmp = business.withColumn('category', explode(split(business['categories'],','))).withColumn('category', trim('category'))
tmp.groupBy('category').count().orderBy('count',ascending=False).show()
#selected_category = ['Restaurants','Food','Coffee & Tea','Sandwiches','Breakfast & Brunch']
regex_expr = r'\b(Restaurants|Food|Coffee|Tea|Sandwiches|Breakfast|Brunch)\b'
business = business.filter(business['categories'].rlike(regex_expr))


business.count()

business.groupBy('city').count().orderBy('count',ascending = False).show()
selected_city = ['Portland']
business = business.filter(col('city').isin(selected_city))


business = business[business['is_open']==1]
business.select([count(when(isnan(c) | col(c).isNull(), c)).alias(c) for c in business.columns]).show()

business.count()

# Review

path = 'yelp_academic_dataset_review.json'
review = spark.read.json(path)
review.count()

review.printSchema()

review.show(10)
review = review.withColumn('year', substring('date',1,4).astype('int'))
review = review.withColumn('month', substring('date',6,2).astype('int'))
review = review.withColumn('day', substring('date',9,2).astype('int'))

review_col = ['review_id','business_id','user_id','year','month','day','stars','useful','funny','cool','text']
review = review.select(review_col)

review.show(10)

review.count()

## Filter Reviews in Bussiness Dataset

business_id_set = set()
for i in list(business.select('business_id').collect()):
    business_id_set.add(i[0])

review.count()

review = review.filter(review['business_id'].isin(business_id_set))

review.count()

## Remove Reviews Before 2010

review.groupBy('year').count().orderBy('year').show()

review = review.filter(review.year>2010)

review.count()

# User

path = 'yelp_academic_dataset_user.json'
user = spark.read.json(path)

user.printSchema()

user_col = ['user_id','name','review_count','yelping_since','friends','useful','funny','cool','fans','average_stars']
user = user.select(user_col)

user.show(10)

# get active user
user = user[user.review_count>10]

user.count()

## Save to CSV

# col = ['business_id','name','city','state','stars','review_count','categories','latitude','longitude','is_open', 'postal_code']
# city = ['Portland']
# category =  ['Restaurants','Food','Coffee & Tea','Sandwiches','Breakfast & Brunch']
# is_open = 1
business.coalesce(1).write.option("header",True).csv("part1_dataclean_business.csv")

# col = ['review_id','business_id','user_id','year','month','day','stars','useful','funny','cool','text']
# only include reviews which "business_id" in "bussiness.json"
review.coalesce(1).write.option("header",True).csv("part1_dataclean_review.csv")

# col = ['user_id','name','review_count','yelping_since','friends','useful','funny','cool','fans','average_stars']
user.coalesce(1).write.option("header",True).csv("part1_dataclean_user.csv")
