==> part5_content_based_filtering.ipynb <==
# Part 5 Content Based Filtering

### Setup

from operator import add
from pyspark.sql import SparkSession
from pyspark.ml.feature import *
from pyspark.ml import Pipeline, PipelineModel
import numpy as np
from pyspark.sql.types import *
from pyspark.sql.functions import *

spark = SparkSession.builder\
    .config("spark.driver.memory", "8g")\
    .config("spark.executor.memory","8g")\
    .config("spark.driver.maxResultSize","0")\
    .config("spark.sql.autoBroadcastJoinThreshold","-1")\
    .config("spark.sql.broadcastTimeout","1200")\
    .config("spark.default.parallelism", "8")\
    .config("spark.network.timeout","7200s")\
    .config("spark.executor.heartbeatInterval", "3600s")\
    .config("spark.local.dir","D:\\Data")\
    .appName("pipeline").getOrCreate()
# spark.local.dir: disable if your do not have this dir

### Read in Pre-processed Data

business = spark.read.csv("part1_dataclean_business.csv",header=True, multiLine=True)
review = spark.read.csv("part4_topicmodeling_review.csv",header=True, multiLine=True)

### Join Review Text by Business ID

reviews_text_rdd = review.select("business_id", "text").rdd
reviews_by_business_rdd = reviews_text_rdd.map(tuple).reduceByKey(add)  
reviews_by_business_df = spark.createDataFrame(reviews_by_business_rdd)
reviews_by_business_df = reviews_by_business_df.withColumnRenamed("_1", "business_id").withColumnRenamed("_2", "text")

### Build Content Based Filtering Pipeline

"""
# prepare the pipeline

regex_tokenizer = RegexTokenizer(gaps = False, pattern = "\w+", inputCol = "text", outputCol = "token")
stop_words_remover = StopWordsRemover(inputCol = "token", outputCol = "non_stop_word")
count_vectorizer = CountVectorizer(inputCol="non_stop_word", outputCol="raw_feature")
idf = IDF(inputCol="raw_feature", outputCol="idf_vector")
word_2_vector = Word2Vec(vectorSize = 100, minCount = 5, inputCol = "non_stop_word", outputCol = "word_vector", seed=1)
vector_assembler = VectorAssembler(inputCols=["idf_vector", "word_vector"], outputCol="combined_vector")

# fit the pipeline

pipeline = Pipeline(stages=[regex_tokenizer, stop_words_remover, count_vectorizer, idf, word_2_vector, vector_assembler])
pipeline_model = pipeline.fit(reviews_by_business_df)

# save the pipeline model

pipeline_model.write().overwrite().save("content_based_pipeline_model")
"""

# load the pipeline model

pipeline_model = PipelineModel.load("content_based_pipeline_model")
all_business_vectors_df = pipeline_model.transform(reviews_by_business_df)

all_business_vectors_df.printSchema()
all_business_vectors_df.show(3)

# save all business vectors

all_business_vectors_df.select("business_id", "word_vector").write.mode("overwrite")\
    .format("parquet").save("part5_all_business_vectors.parquet")

# test load all business vectors

all_business_vectors_df = spark.read.load("part5_all_business_vectors.parquet")
all_business_vectors_df.show(3)

all_business_vectors = all_business_vectors_df.rdd.map(lambda x: (x[0], x[1])).collect()
print(all_business_vectors[1][1])

# utility functions

def cosine_similarity(vector1, vector2): 
    return np.dot(vector1, vector2) / np.sqrt(np.dot(vector1, vector1)) / np.sqrt(np.dot(vector2, vector2)) 


def content_based_get_details(recommended_business):
    return recommended_business.join(business, "business_id", "inner").orderBy("score", ascending=False)
            

### Get Similar Business by List of Business IDs

def get_similar_business(business_ids, limit=10):
    schema = StructType([   
                            StructField("business_id", StringType(), True),
                            StructField("score", IntegerType(), True)
                        ])
    similar_businesses_df = spark.createDataFrame([], schema)
    
    for b_id in business_ids:
        try:
            input_vec = [(r[1]) for r in all_business_vectors if r[0] == b_id][0]
        except:
            continue
        similar_business_rdd = spark.sparkContext.parallelize(
            (i[0], float(cosine_similarity(input_vec, i[1]))) for i in all_business_vectors)

        similar_business_df = spark.createDataFrame(similar_business_rdd) \
            .withColumnRenamed("_1", "business_id") \
            .withColumnRenamed("_2", "score") \
            .orderBy("score", ascending = False)
            
        similar_business_df = similar_business_df.filter(col("business_id") != b_id).limit(limit)
        similar_businesses_df = similar_businesses_df \
                                    .union(similar_business_df)
    return similar_businesses_df.orderBy("score", ascending=False).limit(limit)

# test get similar business

bids = ["Npm0cjoyWwyV13OULL9qOA", "ZKPqinA-7gkDA8-yzubSYw"]

print("input restaurants details:")
business.select("business_id","name", "categories") \
    .filter(business.business_id.isin(bids) == True).show(truncate=False)
    
# get top 10 similar business
similar_business = content_based_get_details(get_similar_business(bids))

print("top 10 similar restaurants: ")
similar_business.toPandas()

### Get Similar Business by User ID

def get_user_recommendation(user_id, limit = 10):
    print(f"[content based filtering] user id: {user_id}")
    user_reviewed_businesses = review\
        .filter((col("user_id") == user_id) & (col("superscore")>=3.0))\
        .orderBy("superscore", ascending=False)\
        .select("business_id").distinct().limit(limit).rdd.collect()
    if len(user_reviewed_businesses)==0:
        return None
    business_list = []
    for row in user_reviewed_businesses:
        business_list.append(row.business_id)
    print(f"[content based filtering] user reviewed businesses: {business_list}")
    return content_based_get_details(get_similar_business(business_list, limit))

# test using user id
user = "uLhdaZUsVvT0gbNTdOSzDg"
user_recommendations =get_user_recommendation(user)
    
print("top 10 similar restaurants: ")
user_recommendations.toPandas()

### Get Similar Business by Keywords

def get_keywords_recommendation(keywords, limit=10):
    
    print(f"[content based filtering] keywords: {keywords}")
    keywords_df = spark.sparkContext.parallelize([(0, keywords)]).toDF(["business_id", "text"])
    
    # transform the the keywords to vectors
    
    transformed_keywords_df = pipeline_model.transform(keywords_df)    
    keywords_vector = transformed_keywords_df.select("word_vector").collect()[0][0]
    sim_bus_byword_rdd = spark.sparkContext\
        .parallelize((i[0], float(cosine_similarity(keywords_vector, i[1]))) for i in all_business_vectors)

    sim_bus_byword_df = spark.createDataFrame(sim_bus_byword_rdd) \
         .withColumnRenamed("_1", "business_id") \
         .withColumnRenamed("_2", "score") \
         .orderBy("score", ascending = False) \
         .limit(limit)
    
    # return top 10 similar businesses
    
    a = sim_bus_byword_df.limit(limit)
    return content_based_get_details(a)

# test using keywords

get_keywords_recommendation("chinese noodles").toPandas()

# more test

get_keywords_recommendation("burger").toPandas()
