==> part6_collaborative_filtering.ipynb <==
# Part 6 Collaborative Filtering

### Setup

from pyspark.sql.functions import *
from pyspark.sql.types import *
from pyspark.ml.recommendation import ALS, ALSModel
from pyspark.ml.tuning import ParamGridBuilder, CrossValidator
from pyspark.ml.evaluation import RegressionEvaluator
from pyspark.sql import SparkSession
import pandas as pd

spark = SparkSession.builder\
    .config("spark.driver.memory", "32g")\
    .config("spark.executor.memory","32g")\
    .config("spark.driver.maxResultSize","0")\
    .config("spark.sql.autoBroadcastJoinThreshold","-1")\
    .config("spark.sql.broadcastTimeout","1200")\
    .config("spark.default.parallelism", "32")\
    .config("spark.executor.heartbeatInterval", "3600")\
    .config("spark.local.dir","D:\\Data")\
    .appName("part6").getOrCreate()
# spark.local.dir: disable if your do not have this dir

business = spark.read.csv("part1_dataclean_business.csv", header=True, multiLine=True)
review = spark.read.csv("part4_topicmodeling_review.csv", header=True, multiLine=True)
user = spark.read.csv("part1_dataclean_user.csv", header=True, multiLine=True)

print("business schema")
business.printSchema()

print("review schema")
review.printSchema()

print("user schema")
user.printSchema()

### Prepare Data for ALS

def df_zip_with_index (df, col_name, offset=0):
    new_schema = StructType(
                    [StructField(col_name,LongType(),True)]       
                    + df.schema.fields                           
                )

    zipped_rdd = df.rdd.zipWithIndex()
    new_rdd = zipped_rdd.map(lambda args: ([args[1] + offset] + list(args[0])))
    return spark.createDataFrame(new_rdd, new_schema)

# prepare user

user_new_df = df_zip_with_index(user, "user_index")
user_new_df.show(2)

user_index_df = user_new_df.select("user_id", "user_index")
user_index_df.show(2)

# prepare business

business_new_df = df_zip_with_index(business, "business_index").cache()
business_new_df.write.mode("overwrite")\
    .format("parquet").save("part6_business_with_index.parquet")

business_new_df = spark.read.load("part6_business_with_index.parquet")
business_new_df.show(2)

business_index_df = business_new_df.select("business_id", "business_index")
business_index_df.show(2)

# map new user_id and business_id in the review dataframe

review_df = review.select("user_id", "business_id", "superscore")
review_new_df = review_df.join(user_index_df, "user_id").join(business_index_df,"business_id")
review_new_df.show(2)

### Build ALS Model

# create the rating matrix required by the ALS model

rating_df = review_new_df.select("user_index", "business_index", review_new_df.superscore.cast("float").alias("rating"))
rating_df.show(2)
rating_df.printSchema()
rating_df.limit(2).toPandas()

(train, test) = rating_df.randomSplit([0.8, 0.2], seed=1)

"""
print("start tuning ALS model")

als = ALS(userCol="user_index", itemCol="business_index", ratingCol="rating", coldStartStrategy="drop")
param_grid = ParamGridBuilder()\
    .addGrid(als.rank,[10, 15, 20])\
    .addGrid(als.maxIter,[10, 15, 20])\
    .build()
evaluator = RegressionEvaluator(metricName="rmse",labelCol="rating")

cv = CrossValidator(estimator=als, estimatorParamMaps=param_grid, evaluator=evaluator, numFolds=5, seed=1)
cv_als_model = cv.fit(train)

als_predictions = cv_als_model.bestModel.transform(test)
evaluator = RegressionEvaluator(metricName="rmse", labelCol="rating", predictionCol="prediction")
rmse = evaluator.evaluate(als_predictions)
print("rmse: " + str(rmse))

best_model = cv_als_model.bestModel
#best_rank is 20
best_model.rank
#best_maxIter is 20
best_model._java_obj.parent().getMaxIter()
# rmse is 1.3383
print("end tuning ALS model")

print("rank:", best_model._java_obj.parent().getRank())
print("maxIter:", best_model._java_obj.parent().getMaxIter())
print("regParam:", best_model._java_obj.parent().getRegParam())
"""

"""
print("start building ALS model")

als = ALS(rank=20, maxIter=20, regParam=0.3, userCol="user_index", itemCol="business_index", ratingCol="rating", \
               coldStartStrategy="drop", seed=1)
als_model = als.fit(train)

als_predictions = als_model.transform(test)
# evaluator = RegressionEvaluator(metricName="rmse", labelCol="rating", predictionCol="prediction")
# rmse = evaluator.evaluate(als_predictions)
# print("rmse: " + str(rmse))

# save the ALS model
als_model.write().overwrite().save("als_model")
# rmse is 1.2579

print("end building ALS model")

"""

# load als model
als_model = ALSModel.load("als_model")

# generate top 10 business for each user

user_recommendations = als_model.recommendForAllUsers(10)
user_recommendations.show(2)

# add the column user_id, cache the recommendaton dataframe and show recommedations sample

all_user_recommendations = user_recommendations.join(user_index_df, "user_index", "inner").cache()
all_user_recommendations.show(2, truncate=False)

# save all user recommendations

all_user_recommendations.write.mode("overwrite")\
    .format("parquet").save("part6_all_user_recommendations.parquet")

# test load

loaded_user_recommendations = spark.read.load("part6_all_user_recommendations.parquet").cache()
loaded_user_recommendations.show(3)
loaded_user_recommendations.printSchema()

def get_collaborative_recommendation(user_id, limit=10):

    filtered_business = loaded_user_recommendations.filter(col("user_id") == user_id)
    if not filtered_business.head(1):
        return None
    recommended_business =  spark.createDataFrame(
        filtered_business.rdd.flatMap(lambda p: p[1])
    )
    return business_new_df.join(recommended_business, "business_index", "inner")\
             .orderBy("rating", ascending = False).limit(limit)

    

# test a random case
random_user = loaded_user_recommendations.sample(0.01, seed=1).head(1)[0].user_id
print(f"[collaborative filtering] test user id: {random_user}")
get_collaborative_recommendation(random_user).toPandas()
