==> part8_hybrid_recommendation.ipynb <==
# Part 8 Hybrid Recommendation

from pyspark.ml import PipelineModel
from pyspark.sql.functions import *
from pyspark.sql.types import FloatType,IntegerType
from pyspark.sql import SparkSession
import pandas as pd
from pyspark.sql.types import *
from pyspark.sql.functions import *
from pyspark.ml.clustering import KMeans
from pyspark.ml.clustering import KMeansModel
from pyspark.ml.feature import VectorAssembler
import numpy as np
import folium
import html
import matplotlib.pyplot as plt
import plotly 
import plotly.offline as py
import plotly.graph_objs as go
import plotly_express as px
from random import sample

spark = SparkSession.builder\
    .config("spark.driver.memory", "32g")\
    .config("spark.executor.memory","32g")\
    .config("spark.driver.maxResultSize","0")\
    .config("spark.sql.autoBroadcastJoinThreshold","-1")\
    .config("spark.sql.broadcastTimeout","1200")\
    .config("spark.default.parallelism", "32")\
    .config("spark.executor.heartbeatInterval", "3600")\
    .config("spark.local.dir","D:\\Data")\
    .appName("part8").getOrCreate()

review = spark.read.csv("part4_topicmodeling_review.csv", header=True, multiLine=True).cache()
business = spark.read.csv("part1_dataclean_business.csv",header=True, multiLine=True).cache()
user = spark.read.csv("part1_dataclean_user.csv", header=True, multiLine=True)

# define utility functions

def cosine_similarity(vector1, vector2): 
    return np.dot(vector1, vector2) / np.sqrt(np.dot(vector1, vector1)) / np.sqrt(np.dot(vector2, vector2)) 
def display_in_map(business_df):
    folium_map = folium.Map(location=[45.5, -122.5], zoom_start=10)
    for i, r in business_df.toPandas().iterrows():
        folium.Marker(
                    location =[r.latitude, r.longitude], 
                    popup = html.escape(r["name"]) + '<br>' + 'stars: ' + str(r.stars) + '<br>' + 'reviews count: ' + str(r.review_count),    
                    icon = folium.Icon(color='green')).add_to(folium_map)
    return folium_map

### Prepare Content Based Filtering

# load pipeline model

pipeline_model = PipelineModel.load("content_based_pipeline_model")


# load all business vectors

all_business_vectors_df = spark.read.load("part5_all_business_vectors.parquet").cache()
all_business_vectors_df.show(3)

all_business_vectors = all_business_vectors_df.rdd.map(lambda x: (x[0], x[1])).collect()
print(all_business_vectors[1][1])

all_user_reviewed_businesses = review\
        .filter((col("superscore")>=3.0))\
        .groupBy("user_id")\
        .agg(collect_set("business_id")) \
        .rdd.collectAsMap()
def get_similar_business(business_ids, limit=10):
    schema = StructType([   
                            StructField("business_id", StringType(), True),
                            StructField("score", IntegerType(), True)
                        ])
    similar_businesses_df = spark.createDataFrame([], schema)
    
    for b_id in business_ids:
        try:
            input_vec = [(r[1]) for r in all_business_vectors if r[0] == b_id][0]
        except:
            continue
        similar_business_rdd = spark.sparkContext.parallelize(
            (i[0], float(cosine_similarity(input_vec, i[1]))) for i in all_business_vectors)

        similar_business_df = spark.createDataFrame(similar_business_rdd) \
            .withColumnRenamed("_1", "business_id") \
            .withColumnRenamed("_2", "score") \
            .orderBy("score", ascending = False)
            
        similar_business_df = similar_business_df.filter(col("business_id") != b_id).limit(limit)
        similar_businesses_df = similar_businesses_df \
                                    .union(similar_business_df)
    return similar_businesses_df.orderBy("score", ascending=False).limit(limit)

def get_user_recommendation(user_id, limit = 10):
    print(f"[content based filtering] user id: {user_id}")
    user_reviewed_businesses = all_user_reviewed_businesses.get(user_id)
    if user_reviewed_businesses is None or len(user_reviewed_businesses)==0:
        return None
    print(f"[content based filtering] user reviewed businesses: {user_reviewed_businesses}")
    return content_based_get_details(get_similar_business(user_reviewed_businesses, limit))

def get_keywords_recommendation(keywords, limit=10):
    
    print(f"[content based filtering] keywords: {keywords}")
    keywords_df = spark.sparkContext.parallelize([(0, keywords)]).toDF(["business_id", "text"])
    
    # transform the the keywords to vectors
    
    transformed_keywords_df = pipeline_model.transform(keywords_df)    
    keywords_vector = transformed_keywords_df.select("word_vector").collect()[0][0]
    sim_bus_byword_rdd = spark.sparkContext\
        .parallelize((i[0], float(cosine_similarity(keywords_vector, i[1]))) for i in all_business_vectors)

    similar_business_df = spark.createDataFrame(sim_bus_byword_rdd) \
         .withColumnRenamed("_1", "business_id") \
         .withColumnRenamed("_2", "score") \
         .orderBy("score", ascending = False) \
         .limit(limit)
    
    # return top 10 similar businesses
    
    a = similar_business_df.limit(limit)
    return content_based_get_details(a)

def content_based_get_details(recommended_business):
    return recommended_business.join(business, "business_id", "inner").orderBy("score", ascending=False)

# verify the loading process for content based
# test case 1
similar_business = get_similar_business(["Npm0cjoyWwyV13OULL9qOA", "ZKPqinA-7gkDA8-yzubSYw"])
print(similar_business)

content_based_get_details(similar_business).toPandas()

# test case 2

get_user_recommendation("uLhdaZUsVvT0gbNTdOSzDg").toPandas()

# test case 3

get_keywords_recommendation("burger").toPandas()


### Prepare Collaborative Filtering 

loaded_user_recommendations = spark.read.load("part6_all_user_recommendations.parquet").cache()
loaded_user_recommendations.show(3)
loaded_user_recommendations.printSchema()

business_new_df = spark.read.load("part6_business_with_index.parquet").cache()
business_new_df.show(2)

def get_collaborative_recommendation(user_id, limit=10):

    filtered_business = loaded_user_recommendations.filter(col("user_id") == user_id)
    if not filtered_business.head(1):
        return None
    recommended_business =  spark.createDataFrame(
        filtered_business.rdd.flatMap(lambda p: p[1])
    )
    return business_new_df.join(recommended_business, "business_index", "inner")\
             .orderBy("rating", ascending = False).limit(limit)

random_user = loaded_user_recommendations.sample(0.01, seed=1).head(1)[0].user_id
print(f"[collaborative filtering] test user id: {random_user}")
get_collaborative_recommendation(random_user).toPandas()

### Prepare Friend Recommendation

user_friends = user.select("user_id", "friends").cache()

def get_friends_recommendation(user_id, limit=10):
    friends = user_friends.filter(col("user_id")==user_id).limit(1)
    if friends.count()<1:
        return None
    friend_str = friends.select("friends").rdd.collect()[0][0].replace(" ", "")
    if friend_str=="":
        return None
    friend_list = list(friend_str.split(","))
    friends_businesses = set()
    print(f"[friend recommendation] user's friends: {friend_list}")
    for friend in friend_list:
        reviewed_business = all_user_reviewed_businesses.get(friend)
        if reviewed_business is not None:
            friends_businesses = friends_businesses.union(reviewed_business)
    if len(friends_businesses)==0:
        return None
    if len(friends_businesses) > limit:
        friends_businesses = sample(friends_businesses, limit)
    return get_business_details(friends_businesses) 

def get_business_details(business_list):
    return business.filter(col("business_id").isin(business_list)).orderBy("stars", ascending=False)
get_friends_recommendation("uLhdaZUsVvT0gbNTdOSzDg").toPandas()


### Prepare Location Based Recommendation

# TODO
kmeans_model_path = 'kmean_model'

def read_model(file_path):
    model = KMeansModel.load(kmeans_model_path)
    return model

# get cluster
def get_label(latitude,longtitude,kmeans_model):
    col = ['latitude','longitude']
    vc = VectorAssembler(inputCols=col,outputCol="features")
    new_user = spark.createDataFrame([(latitude,longtitude)],col)
    input_data = vc.transform(new_user)
    res = kmeans_model.transform(input_data)
    return res.select('prediction').collect()[0][0]

# choose top 10 resautant(order by stars) in the target cluster
def location_based_recommand(latitude,longtitude,kmeans_model,pred,need_num):
    cluster = get_label(latitude,longtitude,kmeans_model)
    tmp = pred.filter(pred['prediction'] == cluster)
    return tmp.orderBy(['stars','review_count'],ascending=False).limit(need_num)

# test
def lbrs_test(model,ds,user_id,need_num):
    friends = user_friends.filter(col("user_id")==user_id).limit(1)
    if friends.count()<1:
        need_num = 10
    dataset = ds.withColumn('latitude',ds['latitude'].cast(FloatType())).withColumn('longitude',ds['longitude'].cast(FloatType())).withColumn('review_count',ds['review_count'].cast(IntegerType())).withColumn('stars',ds['stars'].cast(IntegerType()))
    vc = VectorAssembler(inputCols=['latitude','longitude'],outputCol="features")
    dataset = vc.transform(dataset)
    pred = model.transform(dataset)
    test_point = pred.sample(False,0.1).limit(1)
    rec_point = location_based_recommand(test_point.select('latitude','longitude').collect()[0][0],test_point.select('latitude','longitude').collect()[0][1],model,pred,need_num)
    rec_point = rec_point.select("business_id",'name','city','state','stars','review_count','categories','latitude','longitude','is_open','postal_code')
    return rec_point

### Hybrid Recommendation

# final hybrid recommendation model

def get_hybrid_recommendation_by_user(user_id, 
                                      content_based_limit=5, 
                                      collaborative_based_limit=3,
                                      friend_recommendation_limit=2):
    hybrid_result = spark.createDataFrame(spark.sparkContext.emptyRDD(), business.schema)

    print("content based: ")
    content_based_result = get_user_recommendation(user_id, content_based_limit)
    if content_based_result is not None :
        print(content_based_result.toPandas())
        hybrid_result = hybrid_result.union(content_based_result.drop("score"))
    print("end \n\n")
    
    print("collaborative based: ")
    collaborative_result = get_collaborative_recommendation(user_id, collaborative_based_limit)
    if collaborative_result is not None :
        print(collaborative_result.toPandas())
        hybrid_result = hybrid_result.union(collaborative_result.drop("rating", "business_index"))
    print("end \n\n")

    print("friend based: ")
    friend_result = get_friends_recommendation(user_id, friend_recommendation_limit)
    if friend_result is not None :
        print(friend_result.toPandas())
        hybrid_result = hybrid_result.union(friend_result)
    print("end \n\n")
    
    print("hybrid result: ")
    hybrid_result = hybrid_result.dropDuplicates(["name"])
    if hybrid_result is not None :
        print(hybrid_result.toPandas())
    print("end \n\n")

    number_of_result = hybrid_result.count()
    if  number_of_result< 10:
        # TODO add location based result
        # then union
        need_num = 10-number_of_result
        model = read_model(kmeans_model_path)
        lb_res = lbrs_test(model,business,user_id,need_num)
        print(lb_res.toPandas())
        hybrid_result = hybrid_result.union(lb_res)
        print(f"need {10-number_of_result} business from location based")
        pass
    return hybrid_result

# test case 1: valid user id

final_result = get_hybrid_recommendation_by_user("uLhdaZUsVvT0gbNTdOSzDg")
display_in_map(final_result)

# test case 2: invalid or new user id
final_result = get_hybrid_recommendation_by_user("1uLhdaZUsVvT0gbNTdOSzDg")
display_in_map(final_result)
